third_party/blink/web_tests/external/wpt/fetch/api/basic/url-parsing.sub_encoding=x-cp1251-expected.txt

This is a testharness.js-based test.
[FAIL] Response.redirect() uses the UTF-8 URL parser
  assert_equals: expected "%C3%BF" but got "%26%23255%3B"
Harness: the test ran to completion.
